Refactor: reorganize source files into mandatory
